mega update

+++ 82. tickets pro/main.c
@@ -2,7 +2,7 @@
 #define ARR_MAX 1024
 #include <stdio.h>
 
-void get_sums(unsigned long long sums[], unsigned long long n){
+void get_sums(unsigned long long sums[], int n){
 	if (n == 1){
 		for (int i = 0; i <= 9; i++){
 			sums[i] = 1;
@@ -11,8 +11,8 @@ void get_sums(unsigned long long sums[], unsigned long long n){
 	}
 	unsigned long long sums2[ARR_MAX] = { 0 };
 	get_sums(sums2, n / 2);
-	for (unsigned long long i = 0; i <= 9 * n / 2; i++){
-		for (unsigned long long j = 0; j <= 9 * n / 2; j++){
+	for (int i = 0; i <= 9 * n / 2; i++){
+		for (int j = 0; j <= 9 * n / 2; j++){
 			sums[i + j] += sums2[i] * sums2[j];
 		}
 	}
@@ -22,7 +22,7 @@ int main(){
 	unsigned long long sums[ARR_MAX] = { 0 };
 	get_sums(sums, 16);
 	double sum = 0;
-	for (unsigned long long i = 0; i <= 16*9; i++){
+	for (int i = 0; i <= 16 * 9; i++){
 		sum += (1.0*sums[i] / 1E16) * (1.0*sums[i] / 1E16);
 	}
 	printf("%lf\n", sum);
